falta dois mutantes

+++ src/main/java/ecommerce/dto/ProdutoComQuantidadeDTO.java
@@ -2,5 +2,5 @@ package ecommerce.dto;
 
 public record ProdutoComQuantidadeDTO(
     Long produtoId,
-    Integer quantidade
+    Long quantidade
 ) {}
